fix: incineration failing if downstream objects marked for removal
The waste incineration command would fail if downstream objects were
marked for removal if an upstream object also was marked for removal.
Example: If a database was marked for removal, and a table within that
database was also marked for removal, the incineration command would fail
when it tried to drop the table.

To prevent this the list of potential downstream objects of a already
marked object will be filtered out of the list presented to the user.
Example: If a potential databases named 'test_db_drp' has a potential
schema named 'test_db_drp.test_schema_drp' and the database is marked for
removal. The schema will not be presented to the user when asked to
choose schemas to remove.